Region and Greyscale options
Using the greyscale and region options on LocationOnScreen reduces the time for the function call to about 1/3 on the calls for the non-bobber icons. Example f3.png from .17 seconds to 0.066. Interact region is smaller than the center region for matching the bobber. Hold checks didn't work constantly well as grey scale but still resulted in about 1/2 the execution time. Either way, it should help on lower end machines and could offset the cost of checking f3.png in the release loop.